refactor: move LRULinkedHashMap into its own file
