+++ libsrc/ffdec_lib/src/com/jpexs/helpers/LRULinkedHashMap.java
@@ -1,0 +1,34 @@
+/*
+ *  Copyright (C) 2010-2026 JPEXS, All rights reserved.
+ * 
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ * 
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library.
+ */
+package com.jpexs.helpers;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+public class LRULinkedHashMap<K, V> extends LinkedHashMap<K, V> {
+    private final int maxEntries;
+
+    public LRULinkedHashMap(int maxEntries) {
+        super(maxEntries + 1, 1.0f, true);
+        this.maxEntries = maxEntries;
+    }
+
+    @Override
+    protected boolean removeEldestEntry(Map.Entry<K, V> eldest) {
+        return size() > maxEntries;
+    }
+}
